Filter out images and annotations
COCO dataset has been reduced (from 100k+ images to 3k images) to the data
used in Disambiguating Visual Verbs (listed in "full_sense_annotations.csv").
Some images are missing: 5 from "train2014", 1 from "val2014"
In the same way, annotations have been reduced only to the ones that
have an image in the filtered dataset and written to a unique csv file
("filtered_annotations.csv"). Probably if only textual representation
will be used, it makes sense to refilter annotations including the ones
that miss from the filtered dataset but are present in the VerSe
dataset.