Improve embedding and fix normalisation bug
Now when a word is not known by the word2vec NN, the full sentence
is passed by only discarding that word. Previously the full sentence
was discarded. By doing that, there is more data available for the
embedding. Moreover, a bug have been fixed, since the mean vector
was not normalised, the dot product was not considered as a real
cosine similarity.